pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-api-python-client>=2.179.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "lxml>=6.0.0",
    "pandas>=2.3.2",
    "python-dateutil>=2.9.0.post0",
    "regex>=2025.7.34",
    "streamlit>=1.48.1",
]
